requirements.txt
# Core dependencies
Pillow>=10.0
numpy>=1.24
requests>=2.28

# Voice cloning TTS (Qwen via DashScope)
# Uses REST API directly — no SDK needed, just `requests`

# Whisper (speech-to-text alignment for captions)
openai-whisper>=20230918

# YouTube footage sourcing
yt-dlp>=2024.01

# Websocket logging suppression (transient dep, pin if needed)
websockets>=12.0
